Extracted out the guts of the confirmation delivery logic into a ConfirmationDeliveryService
